refactor: centralize client profile fetch
